Extract upsert logic into CandidateUpserter service
We need to be able to call the upsert operation from outside a job for the
Schools Experience service (as it requires an immediate response from the CRM
with the candidate GUID in).

Extract the upsert logic into a new service that can be called independently.
