+++ GetIntoTeachingApi/Jobs/UpsertCandidateJob.cs
@@ -11,7 +11,7 @@ namespace GetIntoTeachingApi.Jobs
 {
     public class UpsertCandidateJob : BaseJob
     {
-        private readonly ICrmService _crm;
+        private readonly ICandidateUpserter _upserter;
         private readonly INotifyService _notifyService;
         private readonly IPerformContextAdapter _contextAdapter;
         private readonly IMetricService _metrics;
@@ -20,7 +20,7 @@ namespace GetIntoTeachingApi.Jobs
 
         public UpsertCandidateJob(
             IEnv env,
-            ICrmService crm,
+            ICandidateUpserter upserter,
             INotifyService notifyService,
             IPerformContextAdapter contextAdapter,
             IMetricService metrics,
@@ -28,7 +28,7 @@ namespace GetIntoTeachingApi.Jobs
             IAppSettings appSettings)
             : base(env)
         {
-            _crm = crm;
+            _upserter = upserter;
             _notifyService = notifyService;
             _contextAdapter = contextAdapter;
             _metrics = metrics;
@@ -61,87 +61,13 @@ namespace GetIntoTeachingApi.Jobs
             }
             else
             {
-                var registrations = ClearTeachingEventRegistrations(candidate);
-                var phoneCall = ClearPhoneCall(candidate);
-                SaveCandidate(candidate);
-                SaveTeachingEventRegistrations(registrations, candidate);
-                SavePhoneCall(phoneCall, candidate);
-                IncrementCallbackBookingQuotaNumberOfBookings(phoneCall);
+                _upserter.Upsert(candidate);
 
                 _logger.LogInformation($"UpsertCandidateJob - Succeeded - {candidate.Id}");
             }
 
             var duration = (DateTime.UtcNow - _contextAdapter.GetJobCreatedAt(context)).TotalSeconds;
             _metrics.HangfireJobQueueDuration.WithLabels(new[] { "UpsertCandidateJob" }).Observe(duration);
-        }
-
-        private void SaveCandidate(Candidate candidate)
-        {
-            _crm.Save(candidate);
-        }
-
-        private IEnumerable<TeachingEventRegistration> ClearTeachingEventRegistrations(Candidate candidate)
-        {
-            // Due to reasons unknown the event registrations relationship can't be deep-inserted
-            // in the same way we do for other relationships - we need to explicitly save them against
-            // the candidate instead.
-            var teachingEventRegistrations = new List<TeachingEventRegistration>(candidate.TeachingEventRegistrations);
-            candidate.TeachingEventRegistrations.Clear();
-            return teachingEventRegistrations;
-        }
-
-        private PhoneCall ClearPhoneCall(Candidate candidate)
-        {
-            if (candidate.PhoneCall == null)
-            {
-                return null;
-            }
-
-            // Due to reasons unknown the phone call relationship can't be deep-inserted
-            // in the same way we do for other relationships - we need to explicitly save them against
-            // the candidate instead.
-            var phoneCall = candidate.PhoneCall;
-            candidate.PhoneCall = null;
-            return phoneCall;
-        }
-
-        private void SaveTeachingEventRegistrations(IEnumerable<TeachingEventRegistration> registrations, Candidate candidate)
-        {
-            foreach (var registration in registrations)
-            {
-                registration.CandidateId = (Guid)candidate.Id;
-                _crm.Save(registration);
-            }
-        }
-
-        private void IncrementCallbackBookingQuotaNumberOfBookings(PhoneCall phoneCall)
-        {
-            if (phoneCall == null)
-            {
-                return;
-            }
-
-            var quota = _crm.GetCallbackBookingQuota(phoneCall.ScheduledAt);
-
-            if (quota == null || !quota.IsAvailable)
-            {
-                return;
-            }
-
-            quota.NumberOfBookings += 1;
-
-            _crm.Save(quota);
-        }
-
-        private void SavePhoneCall(PhoneCall phoneCall, Candidate candidate)
-        {
-            if (phoneCall == null)
-            {
-                return;
-            }
-
-            phoneCall.CandidateId = candidate.Id.ToString();
-            _crm.Save(phoneCall);
         }
     }
 }
