+++ GetIntoTeachingApi/Services/ICandidateUpserter.cs
@@ -1,0 +1,9 @@
+﻿using GetIntoTeachingApi.Models;
+
+namespace GetIntoTeachingApi.Services
+{
+    public interface ICandidateUpserter
+    {
+        public void Upsert(Candidate candidate);
+    }
+}
